First Commit
Everything useful is in ui.py for now. Uses PyQt5 to create a window that takes in images. Needs to have venv installed via ">source venv/scripts/activate" or something in git bash, as well as ">pip install PyQt5". For some reason setText() doesn't like to work but it takes widgets perfectly fine.  
To Do: Make it pretty, find a way to set the text somewhere; Size the image according to the window, not the other way around; output message after file upload